+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Dodo x.x.x (not yet released)
-* Update to IDLeDOM 0.7.1.
+* Update to IDLeDOM 0.7.2.
 * Fix doctype creation and "significant whitespace" handling in DOMParser.
 * Ensure that Document::documentElement is always populated.
 * Ensure that Document::documentElement matches FilteredElementList where
@@ -7,8 +7,9 @@
 * EXPERIMENTAL: Added Node::getExtensionData() and Node::setExtensionData()
   methods to allow end-users to associate additional non-spec data off of
   Nodes.
-* Implement DocumentFragment::querySelector() and ::querySelectorAll()
-* Implement Node::getNodePath() for PHP compatibility
+* Implement DocumentFragment::querySelector() and ::querySelectorAll().
+* Implement Node::getNodePath() for PHP compatibility.
+* Don't export .phan directory in composer package.
 
 # Dodo 0.1.0 (2021-07-04)
 Initial release.